fix(changeset): move duplicated run changeset code to common changeset package
