+++ deployment/common/changeset/run_changeset.go
@@ -1,4 +1,4 @@
-package deployment
+package changeset
 
 import (
 	"fmt"
